==> 2_machine_learning/3_Regresja logistyczna/1_regresja_logistyczna_case_study.ipynb <==
# Regresja logistyczna
Do estymacji modeli będziemy używać pakietu scikit-learn 
https://scikit-learn.org/stable/install.html


#pip install scikit-learn

# Instalacja scikit-learn
#!python -m pip install scikit-learn

# Import wymaganych bibliotek
from  sklearn.linear_model import LogisticRegression
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns

## Zbiór danych 
https://www.kaggle.com/datasets/erdemtaha/cancer-data

Będziemy pracować na zbiorze danych dotyczącym raka piersi.

### Problem biznesowy:
Zadanie będzie polegać na zaprognozowaniu czy dana pacjentka ma raka piersi łagodnego czy złośliwego.
Rozwiązanie to powinno wspomóc lekarzy, w celu zmniejszenia umieralności.

# Sprawdzenie ścieżki


# Katalog nadrzędny


# Ponowne sprawdzenie ścieżki


# Pobranie danych


## Eksploracja danych

# Nagłówek


# usunięcie ostatniej zmiennej


# typy danych


Wszystkie zmienne poza targetem (diagnosis) są numeryczne. Scikit-learn, nawet w przypadku zmiennych kategorycznych, wymaga ich przekodowania.

# sprawdzenie braków danych


Chcemy modelować zjawisko wykrywania nowotworu złośliwego, więc jako 1 oznaczymy właśnie ten typ choroby.

# Liczebność kategorii zmiennej celu


# Enkodowanie zmiennej celu


# badanie korelacji


W przypadku zmiennych kategorycznych należy użyć korelacji rang spearmana. Kilka zmiennych ma niską wartość korelacji, jednak jest sporo z korelacją co najmniej 0.7. Sprawdźmy czy korelacja pomiedzy zmiennymi nie jest zbyt wysoka.


# Macierz korelacji


# Korelacja z najistotniejszą zmienną..


# ...i kolejną...


# ...i kolejną...


# ...i jeszcze jedną


# Finalne zmienne w modelu


## Badanie outlierów
Podstawową metodą badania outlierów jest użycie rozstępu międzykwartylowego. Powinniśmy uzyskać te same wartości, które są widoczne na box-plotach. Teoria mówi, że metoda działa dobrze dla rozkładów normalnych. Często jednak nie sprawdza się tego założenia, o ile nie mamy mocno skośnego rozkładu. W powyższych zmiennych histogramy przypominają kształtem rozkład normalny.

Przypomnienie:
$BW = [xmediana − 1.5 × IQR, xmediana + 1.5 × IQR]$, gdzie $IQR = Q3 – Q1$

Outlier to wartość wykraczająca poza obszar BW. Niekiedy wartość 1.5 ustala się na wyższą


# Napiszemy funkcję do znajdowania outlierów


# Sprawdzenie funkcji


# Znalezienie outlierów dla wytypowanych zmiennycf


# wyznaczenie finalnej zmiennej z outlierami


# Liczebność


## Podział zbioru na trening i test
Używamy tylko jednego modelu, więc zbiór walidacyjny nie jest potrzebny. Dodatkowo, zbiór jest niewielki, więc byłoby utrudnione, aby wydzielić 3 zbiory
Do wyznaczenia zbioru treningowego i testowego użyjemy funkcji train_test_split

Dokumentacja: https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.train_test_split.html

from sklearn.model_selection import train_test_split

# definicja X i y


# Funkcja zwraca 4 objekty. W argumentach używamy test_size lub train_size, który najczęsciej jest udziałem. random_state ustala ziarno losowania, aby wyniki było powtarzalne


# Zapisanie zbiorów na potrzeby przyszłych zajęć
train_x.to_csv('data/cancer_train_x.csv')
test_x.to_csv('data/cancer_test_x.csv')
train_y.to_csv('data/cancer_train_y.csv')
test_y.to_csv('data/cancer_test_y.csv')

## Modelowanie
Dokumentacja: https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html

# Stworzenie objektu modelu


# Estymacja 


## Ocena jakości


# predykcje


from sklearn.metrics import confusion_matrix, accuracy_score

# Confusion matrix train


# Confusion matrix test


# Accuracy train


# Accuracy test


Jakość modelu jest bardzo wysoka i stabilna. W kolejnym kroku należy zastanowić się nad celem biznesowym, aby wybrać najlepszą metrykę.
