==> 2_machine_learning/3_Regresja logistyczna/2_interpretacja_modelu.ipynb <==
# Interpretacja modelu

# Import wymaganych bibliotek
from  sklearn.linear_model import LogisticRegression
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns

os.chdir('../')

# Pobranie danych

df = pd.read_csv('data/Cancer_Data.csv')
train_x = pd.read_csv('data/cancer_train_x.csv').iloc[:,1:]
test_x = pd.read_csv('data/cancer_test_x.csv').iloc[:,1:]
train_y = pd.read_csv('data/cancer_train_y.csv')['target']
test_y = pd.read_csv('data/cancer_test_y.csv')['target']

## Estymacja modelu
Regularyzacja jest metodą ograniczającą przeuczenie. Nakładana jest kara na funkcję celu, która minimalizuje nadmierne dopasowanie. 
W modelu regresji logistycznej mamy regularyzację wg parametrów l1 oraz l2. Siła regularyzacji ustalana jest parametrem C, gdzie niższe wartości, dają silniejszą regularyzację. 

Więcej informacji można znaleźć tu: https://medium.com/@rithpansanga/logistic-regression-and-regularization-avoiding-overfitting-and-improving-generalization-e9afdcddd09d

## dodanie zmiennej kategorycznej


# Usunięcie wersji nominalnej zmiennej

# Stworzenie obiektu modelu


# Estymacja 


#Współczynniki modelu


# Nazwy zmiennych


# Statystyki zmiennych x


Nie możemy porównywać pomiędzy sobą współczynników, ponieważ zakres zmiennych jest różnyc. Do porównania należałoby wystandaryzować zmienne

# Interpretacja zmiennej perimeter_worst


# Interpretacja zmiennej kategorycznej


## Interacje i przekształcenia zmiennych
Model regresji liniowej wykrywa jedynie zależności liniowej, dlatego warto testować:
1. przekształcenia zmiennych (np. logarytm, potęga, 1/x,...)
2. Interakcje - iloczyn dwóch zmiennych (mogą też być kategoryczne)


# Iloczyn zmiennej kategorycznej z nominalną


# Stworzenie obiektu modelu


# Estymacja 


#Współczynniki modelu


# Nazwy zmiennych


# Interpretacja interakcji


Interpretacja przekształcenia zmiennej odbyba się w taki sam sposób jak podstawowych zmiennych, zmienia się tylko definicja zmiennej.
