==> 2_machine_learning/3_Regresja logistyczna/3_regresja_wielomianowa.ipynb <==
# Regresja logistyczna wielomianowa

Model regresji logistycznej, dla targetu z więcej niż dwiema kategoriami.

Zbiór danych: https://www.kaggle.com/datasets/atefehmirnaseri/cell-phone-price

# Import wymaganych bibliotek
from  sklearn.linear_model import LogisticRegression
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
os.chdir('../')

df = pd.read_csv('data/CellPhone.csv')

## Problem biznesowy
Model służący do wyceny telefonów komórkowych (przypisanie do jednej z 4 kategorii cenowych). Model może być wykorzystywanych w obrocie telefonami używanymi. Usprawnia wycenę oraz sprawia, że jest ona dokładniejsza, przez co można szybciej znaleźć nowego nabywcę i zoptymalizować zyski.
Cel szczegółowy: zwiększenie zysków, poprzez precyzyjną wycenę wykupu.

## Data preprocessing

pd.set_option('display.max_columns',None)

# podsumowanie


# nagłówek


# info



# Liczebności klas


Kategorie są równoliczne, więc zbiór jest zbilansowany, nie musimy się więc tym martwić.
### Analiza relacji z targetem

# Lista potencjalnych X


# target name


# Analiza rozkładów i boxplot

 

Największe różnicowanie widzimy dla zmiennej 'ram'. Kilka zmiennych wykazują pewnie różnice, szczególnie pomiędzy kategorią 0 i 3 np. 'battery_power', 'talk_time' oraz 'px_width' . 
Część zmiennych ma skończoną liczbę kategorii, więc potraktujemy je jako zmienne kategoryczne


# Stworzenie słownika z liczbą unikalnyc wartości


# Zmienne z liczbą kategorii <=10


# Wyciągnięcie indeksów


# Nazwy zmiennych


# dodanie stalej = 1


# Udzial klas w poszczegolnych wartosciach zmiennych


Zmienne mają bardzo zbliżony udział poszczególnych kategorii cen w danych wartościach. Jedynie możemy wytypować jedną zmienną: 'four_g'.

# Finalne zmienne


### VIF index
Do sprawdzenia współliniowości można użyć indeksu VIF. Bada on wielowymiarową korelację pomiędzy zmiennymi.
Ogólnie VIF powyżej wartości 5 sugeruje wysoką współliniowość i potrzebę usunięcia zmiennej.

Więcej informacji: https://www.geeksforgeeks.org/detecting-multicollinearity-with-vif-python/


#pip install statsmodels

from statsmodels.stats.outliers_influence import variance_inflation_factor

# Liczenie VIF


# Zmienne po wykluczeniu


# Liczenie VIF


## Modelowanie

from sklearn.model_selection import train_test_split

# definicja X i y


# Podział na zbiory treningowy i testowy


# Stworzenie obiektu + estymacja


# predykcje


# Prawdopodobieństwo


from sklearn.metrics import accuracy_score

# Accuracy train


# Accuracy test
