==> 2_machine_learning/3_Regresja logistyczna/4_ocena_modelu.ipynb <==
# Ocena modelu
## Zbiór danych 
https://www.kaggle.com/datasets/erdemtaha/cancer-data 

Będziemy pracować na zbiorze danych dotyczącym raka piersi.

# Import wymaganych bibliotek
from  sklearn.linear_model import LogisticRegression
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns

# Pobranie danych
os.chdir('../')

df = pd.read_csv('data/Cancer_Data.csv')
train_x = pd.read_csv('data/cancer_train_x.csv').iloc[:,1:]
test_x = pd.read_csv('data/cancer_test_x.csv').iloc[:,1:]
train_y = pd.read_csv('data/cancer_train_y.csv')['target']
test_y = pd.read_csv('data/cancer_test_y.csv')['target']

## Estymacja modelu

# Stworzenie obiektu modelu


# Estymacja 


## Ocena jakości
### Metryki bazujące na klasie

from sklearn.metrics import (f1_score,
                             accuracy_score, 
                             confusion_matrix, 
                             precision_score, 
                             recall_score,
                             classification_report)


# predykcje


# accuracy 


# precision


# recall


# f1 score


# classification_report train


# classification_report test


### Metryki bazujące na prawdopodobieństwie

from sklearn.metrics import roc_auc_score, roc_curve

# Predykcje


# Pole pod krzywą roc


# Przygotwanie danych do wykresu



# Wykres
