==> 2_machine_learning/3_Regresja logistyczna/5_projekt_klasyfikacja.ipynb <==
# Projekt klasyfikacja

Projekt klasyfikacyjny będzie polegać na stworzeniu modelu regresji logistycznej dla analizy churn. 
Cel biznesowy: Zaprognozowanie klientów chcących zrezygnować z usług, w celu szybszej reakcji i przedstawieniu konkurencyjnej oferty, co zmniejszy odsetek odejść.

Kroki do realizacji:
1. Wczytanie i przejrzenie danych.
2. Analiza braków danych.
3. Zbadaj udział odejść dla umów terminowych i bezterminowych (podpowiedź: umowy na czas nieokreślony mają wartość reamining_contract pustą)
4. Stwórz model dla kontraktów terminowych. Wykonaj poniższe kroki:
    - Selekcja zmiennych.
    - Badanie outlierów.
    - Wydzielenie zbioru treningowego i testowego.
    - Stworzenie modelu regresji logistycznej.
    - Ocena modelu.

Zbiór danych: https://www.kaggle.com/datasets/mehmetsabrikunt/internet-service-churn

Zmienne w zbiorze:
- id - unikalny klucz
- is_tv_subscriber - czy klient ma wykupiony pakiet telewizji? (1-tak/0-nie)
- is_movie_package_subscriber - czy klient ma pakiet filmowy? (1-tak/0-nie)
- subscription_age - liczba lat, od kiedy klient ma usługi firmy.
- bill_avg - średnia kwota rachunku z ostatinch 3 miesięcy.
- reamining_contract - ile czasu pozostało do końca umowy. W przypadku wartości pustej, nie ma umowy terminowej.
- service_failure_count - liczba telefonów klienta w celu zgłodzenia awarii.
- download_avg - średnie zużycie internetu w ostatnich 3 miesiącach.
- upload_avg - średnie upload w ostatnich 3 miesiącach.
- download_over_limit - zmienna informująca o pobieraniu powyżej limitu
- churn - 1 oznacza, że klient anulował usługi.

import pandas as pd
import os

# Pobranie danych
os.chdir('../')

df = pd.read_csv('data/internet_service_churn.csv')

df.columns
